* 2 opcode fixes by dave, thx * correced movement packets, server will no longer drop connection when moving * attempted to fix crash in openGL mode in m2 model loader * fixed SCPDatabase::GetFieldByString(), thx dave * added scp files for 3.0.9 client
git-svn-id: https://pseuwow.svn.sourceforge.net/svnroot/pseuwow@347 fda9a0ab-4e26-0410-ba80-894f06a4b443

+++ src/Client/World/Player.h
@@ -149,7 +149,7 @@ public:
 	float _y;
 	float _z;
 	uint32 _guildId;
-	uint8 _flags;
+	uint32 _flags;
 	uint32 _petInfoId;
 	uint32 _petLevel;
 	uint32 _petFamilyId;
